requirements.txt
# general  modules
django >3.2,<3.3
python-decouple
djangorestframework

# third party modules
markdown       
django-filter  
drf-yasg[validation]
rest_framework_simplejwt

# refactoring and test
black
flake8


# deployment modules